refactor: move buildCountrySubdivisonOptions to a separate file
